+++ docs/src/news.md
@@ -4,6 +4,7 @@
 
 - Breaking
     - **miss**: renamed to **findmiss**.
+    - **tabdf**: renamed to **tabd**.
     - **recod_miss** Declare data as missing in a data frame.
 
 - Modifications